pyproject.toml
[project]
name = "vocacional-llm-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "LeTemp2806",email = "sebastianjoserc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "langchain (>=0.3.24,<0.4.0)",
    "chromadb (>=1.0.7,<2.0.0)",
    "llama-cpp-python (>=0.3.8,<0.4.0)",
    "sqlmodel (>=0.0.24,<0.0.25)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "fastapi (==0.115.9)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "langchain-community (>=0.3.23,<0.4.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
alembic = "^1.15.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
